pyproject.toml
[project]
name = "imagine_the_sound"
version = "0.1.0"
description = "Joint_Project"
authors = [
    {name = "Umdoong",email = "umdoongs@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.6"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "poetry-plugin-shell (>=1.0.1,<2.0.0)",
    "django-restframework (>=0.0.1,<0.0.2)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-soft-delete (>=1.0.18,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

